.requirements.txt
Django==4.2
python-dotenv==0.21
djangorestframework==3.13
djangorestframework-simplejwt==5.0
django-cors-headers==3.10
psycopg2==2.9